# Store parser directly instead of extracting document