This gets rid of the std::move(document) by just storing the parser. It obviously requires the parser to be `new`'d up. It definitely doesn't make up the difference and it's even pretty hard to tell if it helps--there's some things better and some things worse. But it could at least eliminate that as a source of confusion.

I'd love to see if other folks get more reproducible results out of this :/

This PR
--------

| filename | filesize (MB) | JSON.parse(ms) | simdjson.lazyParse (ms) | JSON.parse (GB/s) | simdjson.lazyParse (GB/s) | X faster |
| :-- | --: | --: | --: | --: | --: | --: |
| apache_builds.json | 0.13 | 0.516 | 0.345 | 0.25 | 0.37 | 1.50 |
| canada.json | 2.25 | 26.683 | 10.003 | 0.08 | 0.23 | 2.67 |
| citm_catalog.json | 1.73 | 11.995 | 6.859 | 0.14 | 0.25 | 1.75 |
| github_events.json | 0.07 | 0.575 | 0.336 | 0.11 | 0.19 | 1.71 |
| gsoc_2018.json | 3.33 | 20.406 | 10.242 | 0.16 | 0.32 | 1.99 |
| instruments.json | 0.22 | 0.876 | 1.145 | 0.25 | 0.19 | 0.77 |
| marine_ik.json | 2.98 | 21.626 | 38.259 | 0.14 | 0.08 | 0.57 |
| mesh_pretty.json | 1.58 | 7.558 | 18.535 | 0.21 | 0.09 | 0.41 |
| mesh.json | 0.72 | 5.240 | 6.754 | 0.14 | 0.11 | 0.78 |
| numbers.json | 0.15 | 1.072 | 1.289 | 0.14 | 0.12 | 0.83 |
| random.json | 0.51 | 7.396 | 7.477 | 0.07 | 0.07 | 0.99 |
| sf_citylots.json | 189.78 | 2517.173 | 2196.714 | 0.08 | 0.09 | 1.15 |
| twitter.json | 0.63 | 7.515 | 5.221 | 0.08 | 0.12 | 1.44 |
| twitterescaped.json | 0.56 | 2.599 | 4.522 | 0.22 | 0.12 | 0.57 |
| update_center.json | 0.53 | 7.046 | 5.282 | 0.08 | 0.10 | 1.33 |

master
-------

| filename | filesize (MB) | JSON.parse(ms) | simdjson.lazyParse (ms) | JSON.parse (GB/s) | simdjson.lazyParse (GB/s) | X faster |
| :-- | --: | --: | --: | --: | --: | --: |
| apache_builds.json | 0.13 | 0.520 | 0.326 | 0.24 | 0.39 | 1.60 |
| canada.json | 2.25 | 21.146 | 7.168 | 0.11 | 0.31 | 2.95 |
| citm_catalog.json | 1.73 | 9.960 | 6.146 | 0.17 | 0.28 | 1.62 |
| github_events.json | 0.07 | 0.480 | 0.324 | 0.14 | 0.20 | 1.48 |
| gsoc_2018.json | 3.33 | 16.266 | 9.655 | 0.20 | 0.34 | 1.68 |
| instruments.json | 0.22 | 0.707 | 1.104 | 0.31 | 0.20 | 0.64 |
| marine_ik.json | 2.98 | 20.904 | 22.975 | 0.14 | 0.13 | 0.91 |
| mesh_pretty.json | 1.58 | 6.048 | 7.506 | 0.26 | 0.21 | 0.81 |
| mesh.json | 0.72 | 4.118 | 7.391 | 0.18 | 0.10 | 0.56 |
| numbers.json | 0.15 | 0.903 | 1.210 | 0.17 | 0.12 | 0.75 |
| random.json | 0.51 | 6.169 | 5.681 | 0.08 | 0.09 | 1.09 |
| sf_citylots.json | 189.78 | 1886.632 | 1582.920 | 0.10 | 0.12 | 1.19 |
| twitter.json | 0.63 | 10.139 | 3.606 | 0.06 | 0.18 | 2.81 |
| twitterescaped.json | 0.56 | 2.237 | 2.889 | 0.25 | 0.19 | 0.77 |
| update_center.json | 0.53 | 5.976 | 4.634 | 0.09 | 0.12 | 1.29 |