@luizperes 2020-04-16
Hi @jkeiser, 

I checked out your branch (locally) and the performance gets a little worse than before. Here is what the first item of the benchmark looks like:

#### with `std::move`
`apache_builds.json#simdjson x 4,071 ops/sec ±11.20% (65 runs sampled) => 0.246ms`

#### with this PR
`apache_builds.json#simdjson x 2,843 ops/sec ±8.41% (60 runs sampled) => 0.352ms`

#### what we we believe we can achieve
`apache_builds.json#simdjson x 8,578 ops/sec ±0.56% (93 runs sampled) => 0.117ms`

It seems that using a `new` on the parser is even more costly than before?