@jkeiser 2020-04-16
Yeah, it depends what file you're using. Some get worse, some get better. It actually might be real, and I have no explanation. I ran a few more times and threw out the highest results for each, and got this:

This PR
--------

| filename | filesize (MB) | JSON.parse(ms) | simdjson.lazyParse (ms) | JSON.parse (GB/s) | simdjson.lazyParse (GB/s) | X faster |
| :-- | --: | --: | --: | --: | --: | --: |
| apache_builds.json | 0.13 | 0.516 | 0.345 | 0.25 | 0.37 | 1.50 |
| apache_builds.json | 0.13 | 0.520 | 0.345 | 0.24 | 0.37 | 1.51 |
| canada.json | 2.25 | 25.081 | 9.492 | 0.09 | 0.24 | 2.64 |
| canada.json | 2.25 | 24.650 | 10.805 | 0.09 | 0.21 | 2.28 |
| citm_catalog.json | 1.73 | 11.940 | 6.488 | 0.14 | 0.27 | 1.84 |
| citm_catalog.json | 1.73 | 11.588 | 6.507 | 0.15 | 0.27 | 1.78 |
| github_events.json | 0.07 | 0.575 | 0.336 | 0.11 | 0.19 | 1.71 |
| github_events.json | 0.07 | 0.583 | 0.330 | 0.11 | 0.20 | 1.76 |
| gsoc_2018.json | 3.33 | 20.303 | 10.065 | 0.16 | 0.33 | 2.02 |
| gsoc_2018.json | 3.33 | 18.715 | 7.681 | 0.18 | 0.43 | 2.44 |
| instruments.json | 0.22 | 0.876 | 1.145 | 0.25 | 0.19 | 0.77 |
| instruments.json | 0.22 | 0.874 | 1.479 | 0.25 | 0.15 | 0.59 |
| marine_ik.json | 2.98 | 21.626 | 38.259 | 0.14 | 0.08 | 0.57 |
| marine_ik.json | 2.98 | 20.934 | 29.120 | 0.14 | 0.10 | 0.72 |
| mesh_pretty.json | 1.58 | 7.558 | 18.535 | 0.21 | 0.09 | 0.41 |
| mesh_pretty.json | 1.58 | 7.542 | 16.423 | 0.21 | 0.10 | 0.46 |
| mesh.json | 0.72 | 5.240 | 6.754 | 0.14 | 0.11 | 0.78 |
| mesh.json | 0.72 | 4.772 | 7.805 | 0.15 | 0.09 | 0.61 |
| numbers.json | 0.15 | 1.072 | 1.289 | 0.14 | 0.12 | 0.83 |
| numbers.json | 0.15 | 1.107 | 0.887 | 0.14 | 0.17 | 1.25 |
| random.json | 0.51 | 7.396 | 7.477 | 0.07 | 0.07 | 0.99 |
| random.json | 0.51 | 7.192 | 9.362 | 0.07 | 0.05 | 0.77 |
| sf_citylots.json | 189.78 | 2517.173 | 2196.714 | 0.08 | 0.09 | 1.15 |
| sf_citylots.json | 189.78 | 2256.772 | 2363.868 | 0.08 | 0.08 | 0.95 |
| twitter.json | 0.63 | 7.515 | 5.221 | 0.08 | 0.12 | 1.44 |
| twitter.json | 0.63 | 5.951 | 6.967 | 0.11 | 0.09 | 0.85 |
| twitterescaped.json | 0.56 | 2.599 | 4.522 | 0.22 | 0.12 | 0.57 |
| twitterescaped.json | 0.56 | 2.798 | 3.508 | 0.20 | 0.16 | 0.80 |
| update_center.json | 0.53 | 7.046 | 5.282 | 0.08 | 0.10 | 1.33 |
| update_center.json | 0.53 | 7.062 | 5.715 | 0.08 | 0.09 | 1.24 |

master
-------

| filename | filesize (MB) | JSON.parse(ms) | simdjson.lazyParse (ms) | JSON.parse (GB/s) | simdjson.lazyParse (GB/s) | X faster |
| :-- | --: | --: | --: | --: | --: | --: |
| apache_builds.json | 0.13 | 0.525 | 0.276 | 0.24 | 0.46 | 1.91 |
| apache_builds.json | 0.13 | 0.525 | 0.276 | 0.24 | 0.46 | 1.91 |
| canada.json | 2.25 | 24.391 | 8.598 | 0.09 | 0.26 | 2.84 |
| canada.json | 2.25 | 24.651 | 8.195 | 0.09 | 0.27 | 3.01 |
| citm_catalog.json | 1.73 | 11.660 | 7.404 | 0.15 | 0.23 | 1.57 |
| citm_catalog.json | 1.73 | 11.541 | 6.114 | 0.15 | 0.28 | 1.89 |
| github_events.json | 0.07 | 0.569 | 0.287 | 0.11 | 0.23 | 1.98 |
| github_events.json | 0.07 | 0.580 | 0.283 | 0.11 | 0.23 | 2.05 |
| gsoc_2018.json | 3.33 | 22.161 | 10.910 | 0.15 | 0.31 | 2.03 |
| gsoc_2018.json | 3.33 | 17.321 | 7.124 | 0.19 | 0.47 | 2.43 |
| instruments.json | 0.22 | 0.882 | 0.801 | 0.25 | 0.27 | 1.10 |
| instruments.json | 0.22 | 0.849 | 0.825 | 0.26 | 0.27 | 1.03 |
| marine_ik.json | 2.98 | 22.045 | 29.273 | 0.14 | 0.10 | 0.75 |
| marine_ik.json | 2.98 | 22.172 | 29.875 | 0.13 | 0.10 | 0.74 |
| mesh_pretty.json | 1.58 | 6.962 | 9.258 | 0.23 | 0.17 | 0.75 |
| mesh_pretty.json | 1.58 | 7.607 | 8.452 | 0.21 | 0.19 | 0.90 |
| mesh.json | 0.72 | 5.100 | 7.846 | 0.14 | 0.09 | 0.65 |
| mesh.json | 0.72 | 4.780 | 8.469 | 0.15 | 0.09 | 0.56 |
| numbers.json | 0.15 | 1.166 | 1.028 | 0.13 | 0.15 | 1.13 |
| numbers.json | 0.15 | 1.150 | 1.052 | 0.13 | 0.14 | 1.09 |
| random.json | 0.51 | 7.382 | 5.496 | 0.07 | 0.09 | 1.34 |
| random.json | 0.51 | 8.424 | 5.623 | 0.06 | 0.09 | 1.50 |
| sf_citylots.json | 189.78 | 2365.972 | 1848.542 | 0.08 | 0.10 | 1.28 |
| sf_citylots.json | 189.78 | 2458.404 | 1846.338 | 0.08 | 0.10 | 1.33 |
| twitter.json | 0.63 | 5.979 | 3.560 | 0.11 | 0.18 | 1.68 |
| twitter.json | 0.63 | 8.324 | 3.601 | 0.08 | 0.18 | 2.31 |
| twitterescaped.json | 0.56 | 2.922 | 4.038 | 0.19 | 0.14 | 0.72 |
| twitterescaped.json | 0.56 | 2.760 | 2.307 | 0.20 | 0.24 | 1.20 |
| update_center.json | 0.53 | 6.891 | 3.342 | 0.08 | 0.16 | 2.06 |
| update_center.json | 0.53 | 7.567 | 4.585 | 0.07 | 0.12 | 1.65 |