@jkeiser 2020-04-16
Regardless, this is what simdjson_nodejs used to do, and it is the only way to avoid std::move.

I don't recommend checking it in unless we see a more consistent win from it; I really think we want to be stealing the document from simdjson, at *least* to reduce memory pressure.