@jkeiser 2020-04-16
> #### what we we believe we can achieve
> `apache_builds.json#simdjson x 8,578 ops/sec ±0.56% (93 runs sampled) => 0.117ms`

@luizperes where does this come from, BTW? Did simdjson_nodejs used to go this fast?

I went back and looked; [the old code](https://github.com/luizperes/simdjson_nodejs/blob/089fcdf715497f73d32657c3f7847d48e6982c5e/simdjson/bindings.cpp#L174-L178) did std::move on the parser, which this change doesn't even do:

```c++
  ParsedJson *pjh = new ParsedJson(std::move(pj));
  Napi::External<ParsedJson> buffer = Napi::External<ParsedJson>::New(env, pjh,
    [](Napi::Env /*env*/, ParsedJson * data) {
      delete data;
    });
```

I know I must seem like some kind of std::move partisan here, I'm really not. I just suspect we're jumping at it because it's the most obscure thing in the code (which it definitely is!) :)