@luizperes 2020-04-16
> @luizperes where does this come from, BTW? Did simdjson_nodejs used to go this fast?

@jkeiser I know it can (possibly) go that fast (and maybe even a little faster than that, will try to explain below) I got that value from the code below (using `new document())`:

```
Napi::External<dom::document> buffer = Napi::External<dom::document>::New(env, new dom::document(),
    [](Napi::Env /*env*/, dom::document * doc) {
      delete doc;
    });
```

***

simdjson_nodejs has three methods as I think you know (documented [here](https://github.com/luizperes/simdjson_nodejs/blob/master/index.d.ts)): `isValid`, `parse` and `lazyParse`. While `parse` is slow, as explained on #5, `isValid` and `lazyParse` should  have nearly the same cost, because they essentially do the same thing: validate the json, only the `lazyParse` keeps a reference to the document and exposes the fn `valueForKeyPath`.

Also, you see that, up to [here](https://github.com/luizperes/simdjson_nodejs/blob/master/simdjson/bindings.cpp#L134), `lazyParse` is equivalent to `isValid`. When I benchmark the `isValid` instead of `lazyParse`, I get:

#### isValid
`apache_builds.json#simdjson x 8,189 ops/sec ±1.52% (91 runs sampled) => 0.122ms`

Keeping a reference should be very fast, but as there is some extra work (of course), I believe that the cost of doing a `new dom::document()` should be close to our final expected result.

Let me know if you see something wrong in my explanation! :)